requirements.txt
arrow==1.2.1
asgiref==3.5.0
cffi==1.15.0
cryptography==36.0.1
Django==3.0.8
django-fernet-fields==0.6
djangorestframework==3.13.1
Pillow==9.0.0
psycopg2==2.8.6
pycparser==2.21
python-dateutil==2.8.2
pytz==2021.3
reportlab==3.6.5
six==1.16.0
sqlparse==0.4.2
xlwt==1.3.0
